# Implemented ~90% faster matrix invert functions