--- a/src/OpenTK.Mathematics/Matrix/Matrix2.cs
+++ b/src/OpenTK.Mathematics/Matrix/Matrix2.cs
@@ -551,7 +551,7 @@ namespace OpenTK.Mathematics
         /// <exception cref="InvalidOperationException">Thrown if the Matrix2 is singular.</exception>
         public static void Invert(in Matrix2 mat, out Matrix2 result)
         {
-            var det = mat.Determinant;
+            var det = (mat.Row0.X * mat.Row1.Y) - (mat.Row0.Y * mat.Row1.X);
 
             if (det == 0)
             {
@@ -560,10 +560,15 @@ namespace OpenTK.Mathematics
 
             var invDet = 1f / det;
 
+            // Because the c# jit assumes noalias for byref types we need to
+            // save this value as the write to result.Row0.X could change the
+            // value of mat.Row0.X.
+            var row0x = mat.Row0.X;
+
             result.Row0.X = mat.Row1.Y * invDet;
             result.Row0.Y = -mat.Row0.Y * invDet;
             result.Row1.X = -mat.Row1.X * invDet;
-            result.Row1.Y = mat.Row0.X * invDet;
+            result.Row1.Y = row0x * invDet;
         }
 
         /// <summary>