--- a/src/OpenTK.Mathematics/Matrix/Matrix4.cs
+++ b/src/OpenTK.Mathematics/Matrix/Matrix4.cs
@@ -25,8 +25,10 @@ using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+#if NETCOREAPP3_1_OR_GREATER
 using System.Runtime.Intrinsics;
 using System.Runtime.Intrinsics.X86;
+#endif
 
 namespace OpenTK.Mathematics
 {
@@ -1503,6 +1505,7 @@ namespace OpenTK.Mathematics
         /// <exception cref="InvalidOperationException">Thrown if the Matrix4 is singular.</exception>
         public static void Invert(in Matrix4 mat, out Matrix4 result)
         {
+#if NETCOREAPP3_1_OR_GREATER
             if (Sse3.IsSupported)
             {
                 InvertSse3(in mat, out result);
@@ -1511,8 +1514,12 @@ namespace OpenTK.Mathematics
             {
                 InvertFallback(in mat, out result);
             }
+#else
+            InvertFallback(in mat, out result);
+#endif
         }
 
+#if NETCOREAPP3_1_OR_GREATER
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static unsafe void InvertSse3(in Matrix4 mat, out Matrix4 result)
         {
@@ -1724,6 +1731,7 @@ namespace OpenTK.Mathematics
 #pragma warning restore SA1512 // Single-line comments should not be followed by blank lines
 #pragma warning restore SA1515 // Single-line comment should be preceded by blank line
         }
+#endif
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static unsafe void InvertFallback(in Matrix4 mat, out Matrix4 result)