### Purpose of this PR

Optimize the `Matrix4.Invert` and `Matrix3.Invert` functions (and their `double` equivalents).
Also make the `Matrix2.Invert(in Matrix2 mat, out Matrix2 result)` alias safe at the cost of a little performance.

### Testing status

These functions result in either the exact same result as the current code or in the exact same result as their `System.Numerics` counterpart results in.
They also throw exceptions in the same cases and the original function and have correct results even when the references alias.

### Comments

Here is a breakdown of the profiling done for these tests:
Here is the benchmarks used:
[Benchmarks.zip](https://github.com/opentk/opentk/files/6091326/Benchmarks.zip)

Here are the disassembles provided for refrence:
[Benchmarks Disassembly Reports.zip](https://github.com/opentk/opentk/files/6091319/Benchmarks.Disassembly.Reports.zip)

Matrix4:
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.804 (2004/?/20H1)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.200
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  Job-JPYRWY : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Runtime=.NET Core 5.0  RunStrategy=Throughput  

```
|           Method |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|----------------- |----------:|---------:|---------:|------:|-------:|------:|------:|----------:|----------:|
|   OriginalInvert | 486.71 ns | 5.072 ns | 4.744 ns |  1.00 | 0.0706 |     - |     - |     224 B |    3755 B |
| InvertStackalloc | 198.29 ns | 2.197 ns | 1.948 ns |  0.41 |      - |     - |     - |         - |    1345 B |
|  InvertSysNumEqv |  90.04 ns | 0.491 ns | 0.459 ns |  0.19 |      - |     - |     - |         - |    2778 B |
|     InvertSysNum |  26.00 ns | 0.322 ns | 0.286 ns |  0.05 |      - |     - |     - |         - |    1312 B |
|        InvertGLU |  91.10 ns | 1.844 ns | 3.181 ns |  0.19 |      - |     - |     - |         - |    2382 B |
|  InvertUseSysNum |  30.25 ns | 0.525 ns | 0.465 ns |  0.06 |      - |     - |     - |         - |     229 B |
|       InvertSse3 |  16.19 ns | 0.316 ns | 0.364 ns |  0.03 |      - |     - |     - |         - |     666 B |
|        InvertAvx |  15.89 ns | 0.348 ns | 0.477 ns |  0.03 |      - |     - |     - |         - |     494 B |

Matrix3:
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.804 (2004/?/20H1)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.200
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  Job-QDUIUB : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Runtime=.NET Core 5.0  RunStrategy=Throughput  

```
|           Method |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|----------------- |----------:|---------:|---------:|------:|-------:|------:|------:|----------:|----------:|
|   OriginalInvert | 274.61 ns | 5.397 ns | 5.542 ns |  1.00 | 0.0634 |     - |     - |     200 B |    2927 B |
| InvertStackalloc |  92.40 ns | 1.863 ns | 2.844 ns |  0.34 |      - |     - |     - |         - |    1161 B |
|  InvertSysNum4x4 |  25.37 ns | 0.538 ns | 0.869 ns |  0.09 |      - |     - |     - |         - |    1303 B |
|        InvertSMI |  13.76 ns | 0.313 ns | 0.439 ns |  0.05 |      - |     - |     - |         - |     505 B |
|    InvertSMISafe |  15.48 ns | 0.271 ns | 0.290 ns |  0.06 |      - |     - |     - |         - |     663 B |

Matrix2:
``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.804 (2004/?/20H1)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT  [AttachedDebugger]
  Job-BHCUYE : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT

Runtime=.NET Core 5.0  RunStrategy=Throughput  

```
|               Method |     Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|--------------------- |---------:|----------:|----------:|------:|--------:|----------:|
|       OriginalInvert | 7.712 ns | 0.1234 ns | 0.1154 ns |  1.00 |    0.00 |     206 B |
| OriginalInvertSafe | 9.653 ns | 0.1301 ns | 0.1217 ns |  1.25 |    0.03 |     206 B |
|      InvertSysNum3x2 | 5.352 ns | 0.1137 ns | 0.1008 ns |  0.69 |    0.01 |     353 B |